+++ src/nextjournal/clerk/builder.clj
@@ -389,7 +389,8 @@
                       (report-fn {:stage :building :doc doc :idx idx})
                       (let [{result :result duration :time-ms} (eval/time-ms
                                                                 (try
-                                                                  (binding [*build-opts* opts
+                                                                  (binding [*ns* *ns*
+                                                                            *build-opts* opts
                                                                             viewer/doc-url (partial doc-url opts state file)]
                                                                     (let [doc (eval/eval-analyzed-doc doc)]
                                                                       (assoc doc :viewer (view/doc->viewer (assoc opts :static-build? true